pyproject.toml
[tool.poetry]
name = "ppp-connectors"
packages = [{ include = "ppp_connectors" }]
version = "0.1.6"
description = "A simple, lightweight set of connectors and functions to various APIs, controlled by a central broker."
authors = [
    "Rob D'Aveta <rob.daveta@gmail.com>",
    "Aaron Buchanan <aaronbuchanan@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
